BackEnd/requirements.txt
django==3.2.4
requests==2.25.1
djangorestframework==3.12.4
psycopg2==2.9.1